Part 1 of complete overhaul
