+++ constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+    SUMMONER_BY_NAME_VERSION : '1.4',
+    LEAGUE_BY_SUMMONER_VERSION : '2.5',
+    RECENT_GAME_VERSION : '2.2'
+}
