--- views/base.jade
@@ -1,58 +1,0 @@
-html
-  
-  head
-    title LolDecay
-    link(rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.1.3/semantic.min.css")
-    
-  body
-  
-    br
-    br
-    br
-    br
-    div(class="ui page grid")
-      div(class="center aligned starter column")
-        h1(class="ui header" style="font-size:40px;") LoL Decay
-    
-    block content
-
-    div(class="ui bottom fixed large inverted menu" style="background-color:#1B1C1D")
-        div(class="ui centered grid container")
-          div(class="ui secondary large menu")
-            a(class="item" href="/" style="font-size:20px") Home
-            a(class="item" href="changelog" style="font-size:20px") Changelog
-            a(class="item" href="notify" style="font-size:20px") Notify
-            a(class="item" href="about" style="font-size:20px") About
-            
-
-    script(src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js")
-    script(src="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.1.3/semantic.min.js")
-    script(type="text/javascript") $('.ui.dropdown').dropdown();
-    script(type="text/javascript").  
-      var formValidationRules = {
-        fields: {
-            name: {
-              identifier : 'register_name',
-              rules: [
-                {
-                  type   : 'empty',
-                  prompt : 'Please enter a summoner name!'
-                }
-              ]
-            },
-            email: {
-              identifier : 'register_email',
-              rules: [
-                {
-                  type   : 'email',
-                  prompt : 'Please enter a valid email!'
-                }
-              ]
-            }
-            
-            
-          }
-      };
-          
-      $('.ui.form')
-        .form(formValidationRules);
