--- views/verify.jade
@@ -1,20 +1,0 @@
-extends ./infopages.jade
-
-block title
-  h1(class="ui header" style="font-size:30px;") Notify
-
-block message 
-  
-  a(href="notify" style="font-size:20px") Back to form
-  br
-  
-  h3!= o_msg
-  br
-  
-  if (success) 
-    p(style="font-size:18") Successfully linked #[span(style="font-weight:bold")!= m1] with #[span(style="font-weight:bold")!= m2]
-        
-    p(style="font-size:18") You will receive a confirmation email soon.
-    br
-    p(style="font-size:18") To cancel alerts, see link on incoming email(s). 
-    p(style="font-size:18") Note that your email will not be used for any other purpose besides intended.
