+++ views_old/delete.jade
@@ -1,0 +1,11 @@
+extends ./infopages.jade
+
+block title
+  h1(class="ui header" style="font-size:30px;") Deletion
+
+block message 
+  
+  a(href="/notify" style="font-size:20px") Back to form
+  br
+  br
+  p(style="font-size:18") You have been unsubscribed for this link. Note that you will continue receiving emails for associations with other accounts under your current email.
