+++ views_old/infopages.jade
@@ -1,0 +1,11 @@
+extends ./base.jade
+
+block content
+  div(class="ui center aligned page grid")
+    div(class="row")
+      div(class="five wide center column")
+        block title
+    div(class="row")
+      div(class="twelve wide left aligned column")
+        block message
+
